Create base referral component with shared logic
