+++ embc-app/ClientApp/src/app/volunteer/components/incidentals-referral/incidentals-referral.component.ts
@@ -4,21 +4,18 @@ import { IncidentalsReferral, Evacuee } from 'src/app/core/models';
 import { IncidentalsRatesComponent } from 'src/app/shared/modals/incidentals-rates/incidentals-rates.component';
 import { numberOfDays, uuid, clearFormArray } from 'src/app/shared/utils';
 import { SupplierComponent } from '../supplier/supplier.component';
-import { FormGroup, FormBuilder, Validators, FormArray, FormControl } from '@angular/forms';
+import { FormBuilder, Validators, FormArray, FormControl } from '@angular/forms';
 import { CustomValidators } from 'src/app/shared/validation/custom.validators';
+import { AbstractReferralComponent } from '../abstract-referral/abstract-referral.component';
 
 @Component({
   selector: 'app-incidentals-referral',
   templateUrl: './incidentals-referral.component.html',
   styleUrls: ['./incidentals-referral.component.scss']
 })
-export class IncidentalsReferralComponent implements OnInit, AfterViewInit, OnDestroy, OnChanges {
-  // List of all evacuees that we want to show in this component
-  @Input() evacuees: Evacuee[];
-  @Input() referral: IncidentalsReferral = null;
-  @Input() readOnly = false;
+export class IncidentalsReferralComponent extends AbstractReferralComponent implements OnInit, AfterViewInit, OnDestroy, OnChanges {
 
-  @Output() validityChange = new EventEmitter<boolean>();
+  @Input() referral: IncidentalsReferral = null;
 
   @ViewChild(SupplierComponent) supplier: SupplierComponent;
 
@@ -28,23 +25,17 @@ export class IncidentalsReferralComponent implements OnInit, AfterViewInit, OnDe
   // If it breaks and isn't unique it won't break the form. (poor man's guid)
   uuid = uuid();
 
-  // The model for the form data collected
-  form = this.fb.group({
-    selectedEvacuees: this.fb.array([], Validators.required),
-    approvedItems: '',
-    totalAmount: ['', [CustomValidators.number, Validators.required, Validators.min(0)]],
-  });
-
   constructor(
-    private modals: NgbModal,
-    private fb: FormBuilder,
-  ) { }
-
-  get array() {
-    return this.form.get('selectedEvacuees') as FormArray;
+    public fb: FormBuilder,
+    public modals: NgbModal,
+  ) {
+    super(fb);
+    this.form.addControl('approvedItems', this.fb.control(''));
+    this.form.addControl('totalAmount', this.fb.control('', [CustomValidators.number, Validators.required, Validators.min(0)]));
   }
 
   ngOnInit() {
+    super.ngOnInit();
     this.handleFormChange();
     this.displayReferral(this.referral);
   }
@@ -72,7 +63,6 @@ export class IncidentalsReferralComponent implements OnInit, AfterViewInit, OnDe
   // validate the whole form as we capture data
   handleFormChange(): void {
     this.form.valueChanges.subscribe(() => this.validate());
-    this.form.statusChanges.subscribe(status => this.validityChange.emit(status === 'VALID'));
   }
 
   displayReferral(referral: IncidentalsReferral) {
@@ -101,33 +91,7 @@ export class IncidentalsReferralComponent implements OnInit, AfterViewInit, OnDe
     );
   }
 
-  addEvacuee(evacuee: Evacuee) {
-    this.array.push(new FormControl(evacuee));
-  }
-
-  removeEvacuee(index: number) {
-    this.array.removeAt(index);
-  }
-
-  selectEvacuee(evacuee: Evacuee) {
-    const index = this.indexOfEvacuee(this.array.value, evacuee);
-    if (index >= 0) {
-      this.removeEvacuee(index);
-    } else {
-      this.addEvacuee(evacuee);
-    }
-  }
-
-  selectAllEvacuees() {
-    clearFormArray(this.array);
-    (this.evacuees || []).forEach(x => this.addEvacuee(x));
-  }
-
   // --------------------HELPERS-----------------------------------------
   numDays(validFrom: Date, validTo: Date) { return numberOfDays(validFrom, validTo); }
 
-  // search an Evacuee by ID
-  indexOfEvacuee(arr: Array<Evacuee>, value: Evacuee): number {
-    return (arr || []).map(o => o.id).indexOf(value.id);
-  }
 }
